llm_service/requirements.txt
flask
llama-cpp-python>=0.2.11
